增加 用户 司机 的登录逻辑 及 积分更新

+++ User/src/main/java/distributed/sys/customer/entity/Customer.java
@@ -27,7 +27,7 @@ public class Customer {
     @JsonView(Views.Public.class)
     private int takeCount;
     @JsonView(Views.Public.class)
-    private double takeDistance;
+    private int takeDistance;
 
     @JsonView(Views.Public.class)
     private int membershipPoint;
